+++ src/main/java/com/klu/ss/repository/LikeRepo.java
@@ -4,7 +4,7 @@ import com.klu.ss.model.*;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
-public interface LikeRepo extends JpaRepository<Like, Integer>{
+public interface LikeRepo extends JpaRepository<Like, Long>{
 	 Optional<Like> findByPostAndProfile(Post post, Profile profile);
 	    void deleteByPostAndProfile(Post post, Profile profile);
