+++ src/main/java/com/klu/ss/service/FoodOfferService.java
@@ -7,12 +7,12 @@ import com.klu.ss.model.FoodOffer;
 import com.klu.ss.model.Requesting;
 public interface FoodOfferService {
     FoodOffer saveFoodOffer(FoodOfferDTO foodOfferDTO);
-    Optional<FoodOffer> getFoodOfferById(int oid);
-    List<FoodOffer> getOffersByUserId(int userId);
-    List<Requesting> getRequestsForOffer(int offerId);
+    Optional<FoodOffer> getFoodOfferById(long oid);
+    List<FoodOffer> getOffersByUserId(long userId);
+    List<Requesting> getRequestsForOffer(long offerId);
     boolean confirmRequest(Long offerId, Long requesterId);
     boolean markAsCompleted(Long offerId, Long requesterId);
     List<FoodOffer> getAllDonations();
-    List<FoodOffer> getDonationsByUser(int userId);
+    List<FoodOffer> getDonationsByUser(long userId);
 }
