+++ src/main/java/com/klu/ss/service/ProfileService.java
@@ -4,8 +4,8 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.klu.ss.model.Profile;
 public interface ProfileService {
-	public String updateUserDetails(Integer uid, MultipartFile profilePic, MultipartFile bannerPic, String name,
+	public String updateUserDetails(Long uid, MultipartFile profilePic, MultipartFile bannerPic, String name,
 			LocalDate dateOfBirth, String phone, String address, String badge);
-	public Profile getProfileByUserId(int userId);
+	public Profile getProfileByUserId(long userId);
 	public Profile getProfileDetailsByUsername(String username);
 }
